Network.AWS.CognitoIdentityProvider.AddCustomAttributes

Description

Adds additional user attributes to the user pool schema.

Creating a Request

addCustomAttributes

Arguments

:: Text
acaUserPoolId

-> NonEmpty SchemaAttributeType
acaCustomAttributes

-> AddCustomAttributes

Creates a value of AddCustomAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acaUserPoolId - The user pool ID for the user pool where you want to add custom attributes.

acaCustomAttributes - An array of custom attributes, such as Mutable and Name.

data AddCustomAttributes

Represents the request to add custom attributes.

See: addCustomAttributes smart constructor.

Request Lenses

acaUserPoolId :: Lens' AddCustomAttributes Text

The user pool ID for the user pool where you want to add custom attributes.

acaCustomAttributes :: Lens' AddCustomAttributes (NonEmpty SchemaAttributeType)

An array of custom attributes, such as Mutable and Name.

Destructuring the Response

addCustomAttributesResponse

Arguments

:: Int
acarsResponseStatus

-> AddCustomAttributesResponse

Creates a value of AddCustomAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acarsResponseStatus - -- | The response status code.

data AddCustomAttributesResponse

Represents the response from the server for the request to add custom attributes.

See: addCustomAttributesResponse smart constructor.

Response Lenses

acarsResponseStatus :: Lens' AddCustomAttributesResponse Int

- | The response status code.