Network.AWS.CognitoIdentityProvider.AdminConfirmSignUp

Description

Confirms user registration as an admin without using a confirmation code. Works on any user.

Requires developer credentials.

Creating a Request

adminConfirmSignUp

Arguments

:: Text
acsuUserPoolId

-> Text
acsuUsername

-> AdminConfirmSignUp

Creates a value of AdminConfirmSignUp with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acsuUserPoolId - The user pool ID for which you want to confirm user registration.

acsuUsername - The user name for which you want to confirm user registration.

data AdminConfirmSignUp

Represents the request to confirm user registration.

See: adminConfirmSignUp smart constructor.

Request Lenses

acsuUserPoolId :: Lens' AdminConfirmSignUp Text

The user pool ID for which you want to confirm user registration.

acsuUsername :: Lens' AdminConfirmSignUp Text

The user name for which you want to confirm user registration.

Destructuring the Response

adminConfirmSignUpResponse

Arguments

:: Int
acsursResponseStatus

-> AdminConfirmSignUpResponse

Creates a value of AdminConfirmSignUpResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acsursResponseStatus - -- | The response status code.

data AdminConfirmSignUpResponse

Represents the response from the server for the request to confirm registration.

See: adminConfirmSignUpResponse smart constructor.

Response Lenses

acsursResponseStatus :: Lens' AdminConfirmSignUpResponse Int

- | The response status code.