Network.AWS.CognitoIdentityProvider.AdminCreateUser

Description

Creates a new user in the specified user pool and sends a welcome message via email or phone (SMS). This message is based on a template that you configured in your call to CreateUserPool or UpdateUserPool . This template includes your custom sign-up instructions and placeholders for user name and temporary password.

Requires developer credentials.

Creating a Request

adminCreateUser

Arguments

:: Text
acuUserPoolId

-> Text
acuUsername

-> AdminCreateUser

Creates a value of AdminCreateUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acuTemporaryPassword - The user's temporary password. This password must conform to the password policy that you specified when you created the user pool. The temporary password is valid only once. To complete the Admin Create User flow, the user must enter the temporary password in the sign-in page along with a new password to be used in all future sign-ins. This parameter is not required. If you do not specify a value, Amazon Cognito generates one for you. The temporary password can only be used until the user account expiration limit that you specified when you created the user pool. To reset the account after that time limit, you must call AdminCreateUser again, specifying RESEND for the MessageAction parameter.

acuForceAliasCreation - This parameter is only used if the phone_number_verified or email_verified attribute is set to True . Otherwise, it is ignored. If this parameter is set to True and the phone number or email address specified in the UserAttributes parameter already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user. The previous user will no longer be able to log in using that alias. If this parameter is set to False , the API throws an AliasExistsException error if the alias already exists. The default value is False .

acuDesiredDeliveryMediums - Specify EMAIL if email will be used to send the welcome message. Specify SMS if the phone number will be used. The default value is SMS . More than one value can be specified.

acuMessageAction - Set to RESEND to resend the invitation message to a user that already exists and reset the expiration limit on the user's account. Set to SUPPRESS to suppress sending the message. Only one value can be specified.

acuUserAttributes - An array of name-value pairs that contain user attributes and attribute values to be set for the user to be created. You can create a user without specifying any attributes other than Username . However, any attributes that you specify as required (in CreateUserPool or in the Attributes tab of the console) must be supplied either by you (in your call to AdminCreateUser ) or by the user (when he or she signs up in response to your welcome message). For custom attributes, you must prepend the custom: prefix to the attribute name. To send a message inviting the user to sign up, you must specify the user's email address or phone number. This can be done in your call to AdminCreateUser or in the Users tab of the Amazon Cognito console for managing your user pools. In your call to AdminCreateUser , you can set the email_verified attribute to True , and you can set the phone_number_verified attribute to True . (You can also do this by calling AdminUpdateUserAttributes .) * email : The email address of the user to whom the message that contains the code and username will be sent. Required if the email_verified attribute is set to True , or if EMAIL is specified in the DesiredDeliveryMediums parameter. * phone_number : The phone number of the user to whom the message that contains the code and username will be sent. Required if the phone_number_verified attribute is set to True , or if SMS is specified in the DesiredDeliveryMediums parameter.

acuValidationData - The user's validation data. This is an array of name-value pairs that contain user attributes and attribute values that you can use for custom validation, such as restricting the types of user accounts that can be registered. For example, you might choose to allow or disallow user sign-up based on the user's domain. To configure custom validation, you must create a Pre Sign-up Lambda trigger for the user pool as described in the Amazon Cognito Developer Guide. The Lambda trigger receives the validation data and uses it in the validation process. The user's validation data is not persisted.

acuUserPoolId - The user pool ID for the user pool where the user will be created.

acuUsername - The username for the user. Must be unique within the user pool. Must be a UTF-8 string between 1 and 128 characters. After the user is created, the username cannot be changed.

data AdminCreateUser

Represents the request to create a user in the specified user pool.

See: adminCreateUser smart constructor.

Request Lenses

acuTemporaryPassword :: Lens' AdminCreateUser (Maybe Text)

The user's temporary password. This password must conform to the password policy that you specified when you created the user pool. The temporary password is valid only once. To complete the Admin Create User flow, the user must enter the temporary password in the sign-in page along with a new password to be used in all future sign-ins. This parameter is not required. If you do not specify a value, Amazon Cognito generates one for you. The temporary password can only be used until the user account expiration limit that you specified when you created the user pool. To reset the account after that time limit, you must call AdminCreateUser again, specifying RESEND for the MessageAction parameter.

acuForceAliasCreation :: Lens' AdminCreateUser (Maybe Bool)

This parameter is only used if the phone_number_verified or email_verified attribute is set to True . Otherwise, it is ignored. If this parameter is set to True and the phone number or email address specified in the UserAttributes parameter already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user. The previous user will no longer be able to log in using that alias. If this parameter is set to False , the API throws an AliasExistsException error if the alias already exists. The default value is False .

acuDesiredDeliveryMediums :: Lens' AdminCreateUser [DeliveryMediumType]

Specify EMAIL if email will be used to send the welcome message. Specify SMS if the phone number will be used. The default value is SMS . More than one value can be specified.

acuMessageAction :: Lens' AdminCreateUser (Maybe MessageActionType)

Set to RESEND to resend the invitation message to a user that already exists and reset the expiration limit on the user's account. Set to SUPPRESS to suppress sending the message. Only one value can be specified.

acuUserAttributes :: Lens' AdminCreateUser [AttributeType]

An array of name-value pairs that contain user attributes and attribute values to be set for the user to be created. You can create a user without specifying any attributes other than Username . However, any attributes that you specify as required (in CreateUserPool or in the Attributes tab of the console) must be supplied either by you (in your call to AdminCreateUser ) or by the user (when he or she signs up in response to your welcome message). For custom attributes, you must prepend the custom: prefix to the attribute name. To send a message inviting the user to sign up, you must specify the user's email address or phone number. This can be done in your call to AdminCreateUser or in the Users tab of the Amazon Cognito console for managing your user pools. In your call to AdminCreateUser , you can set the email_verified attribute to True , and you can set the phone_number_verified attribute to True . (You can also do this by calling AdminUpdateUserAttributes .) * email : The email address of the user to whom the message that contains the code and username will be sent. Required if the email_verified attribute is set to True , or if EMAIL is specified in the DesiredDeliveryMediums parameter. * phone_number : The phone number of the user to whom the message that contains the code and username will be sent. Required if the phone_number_verified attribute is set to True , or if SMS is specified in the DesiredDeliveryMediums parameter.

acuValidationData :: Lens' AdminCreateUser [AttributeType]

The user's validation data. This is an array of name-value pairs that contain user attributes and attribute values that you can use for custom validation, such as restricting the types of user accounts that can be registered. For example, you might choose to allow or disallow user sign-up based on the user's domain. To configure custom validation, you must create a Pre Sign-up Lambda trigger for the user pool as described in the Amazon Cognito Developer Guide. The Lambda trigger receives the validation data and uses it in the validation process. The user's validation data is not persisted.

acuUserPoolId :: Lens' AdminCreateUser Text

The user pool ID for the user pool where the user will be created.

acuUsername :: Lens' AdminCreateUser Text

The username for the user. Must be unique within the user pool. Must be a UTF-8 string between 1 and 128 characters. After the user is created, the username cannot be changed.

Destructuring the Response

adminCreateUserResponse

Arguments

:: Int
acursResponseStatus

-> AdminCreateUserResponse

Creates a value of AdminCreateUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

acursUser - The newly created user.

acursResponseStatus - -- | The response status code.

data AdminCreateUserResponse

Represents the response from the server to the request to create the user.

See: adminCreateUserResponse smart constructor.

Response Lenses

acursUser :: Lens' AdminCreateUserResponse (Maybe UserType)

The newly created user.

acursResponseStatus :: Lens' AdminCreateUserResponse Int

- | The response status code.