Network.AWS.CognitoIdentityProvider.AdminDeleteUser

Description

Deletes a user as an administrator. Works on any user.

Requires developer credentials.

Creating a Request

adminDeleteUser

Arguments

:: Text
aUserPoolId

-> Text
aUsername

-> AdminDeleteUser

Creates a value of AdminDeleteUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aUserPoolId - The user pool ID for the user pool where you want to delete the user.

aUsername - The user name of the user you wish to delete.

data AdminDeleteUser

Represents the request to delete a user as an administrator.

See: adminDeleteUser smart constructor.

Request Lenses

aUserPoolId :: Lens' AdminDeleteUser Text

The user pool ID for the user pool where you want to delete the user.

aUsername :: Lens' AdminDeleteUser Text

The user name of the user you wish to delete.

Destructuring the Response

adminDeleteUserResponse :: AdminDeleteUserResponse

Creates a value of AdminDeleteUserResponse with the minimum fields required to make a request.

data AdminDeleteUserResponse

See: adminDeleteUserResponse smart constructor.