Network.AWS.CognitoIdentityProvider.AdminGetUser

Description

Gets the specified user by user name in a user pool as an administrator. Works on any user.

Requires developer credentials.

Creating a Request

adminGetUser

Arguments

:: Text
aguUserPoolId

-> Text
aguUsername

-> AdminGetUser

Creates a value of AdminGetUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aguUserPoolId - The user pool ID for the user pool where you want to get information about the user.

aguUsername - The user name of the user you wish to retrieve.

data AdminGetUser

Represents the request to get the specified user as an administrator.

See: adminGetUser smart constructor.

Request Lenses

aguUserPoolId :: Lens' AdminGetUser Text

The user pool ID for the user pool where you want to get information about the user.

aguUsername :: Lens' AdminGetUser Text

The user name of the user you wish to retrieve.

Destructuring the Response

adminGetUserResponse

Arguments

:: Int
agursResponseStatus

-> Text
agursUsername

-> AdminGetUserResponse

Creates a value of AdminGetUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

agursEnabled - Indicates that the status is enabled.

agursUserStatus - The user status. Can be one of the following: * UNCONFIRMED - User has been created but not confirmed. * CONFIRMED - User has been confirmed. * ARCHIVED - User is no longer active. * COMPROMISED - User is disabled due to a potential security threat. * UNKNOWN - User status is not known.

agursUserAttributes - An array of name-value pairs representing user attributes.

agursUserCreateDate - The date the user was created.

agursMFAOptions - Specifies the options for MFA (e.g., email or phone number).

agursUserLastModifiedDate - The date the user was last modified.

agursResponseStatus - -- | The response status code.

agursUsername - The user name of the user about whom you are receiving information.

data AdminGetUserResponse

Represents the response from the server from the request to get the specified user as an administrator.

See: adminGetUserResponse smart constructor.

Response Lenses

agursEnabled :: Lens' AdminGetUserResponse (Maybe Bool)

Indicates that the status is enabled.

agursUserStatus :: Lens' AdminGetUserResponse (Maybe UserStatusType)

The user status. Can be one of the following: * UNCONFIRMED - User has been created but not confirmed. * CONFIRMED - User has been confirmed. * ARCHIVED - User is no longer active. * COMPROMISED - User is disabled due to a potential security threat. * UNKNOWN - User status is not known.

agursUserAttributes :: Lens' AdminGetUserResponse [AttributeType]

An array of name-value pairs representing user attributes.

agursUserCreateDate :: Lens' AdminGetUserResponse (Maybe UTCTime)

The date the user was created.

agursMFAOptions :: Lens' AdminGetUserResponse [MFAOptionType]

Specifies the options for MFA (e.g., email or phone number).

agursUserLastModifiedDate :: Lens' AdminGetUserResponse (Maybe UTCTime)

The date the user was last modified.

agursResponseStatus :: Lens' AdminGetUserResponse Int

- | The response status code.

agursUsername :: Lens' AdminGetUserResponse Text

The user name of the user about whom you are receiving information.