Network.AWS.CognitoIdentityProvider.AdminLinkProviderForUser

Description

Links an existing user account in a user pool (DestinationUser ) to an identity from an external identity provider (SourceUser ) based on a specified attribute name and value from the external identity provider. This allows you to create a link from the existing user account to an external federated user identity that has not yet been used to sign in, so that the federated user identity can be used to sign in as the existing user account.

For example, if there is an existing user with a username and password, this API links that user to a federated user identity, so that when the federated user identity is used, the user signs in as the existing user account.

Important: Because this API allows a user with an external federated identity to sign in as an existing user in the user pool, it is critical that it only be used with external identity providers and provider attributes that have been trusted by the application owner.

See also AdminDisableProviderForUser .

This action is enabled only for admin access and requires developer credentials.

Creating a Request

adminLinkProviderForUser

Arguments

:: Text
alpfuUserPoolId

-> ProviderUserIdentifierType
alpfuDestinationUser

-> ProviderUserIdentifierType
alpfuSourceUser

-> AdminLinkProviderForUser

Creates a value of AdminLinkProviderForUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alpfuUserPoolId - The user pool ID for the user pool.

alpfuDestinationUser - The existing user in the user pool to be linked to the external identity provider user account. Can be a native (Username + Password) Cognito User Pools user or a federated user (for example, a SAML or Facebook user). If the user doesn't exist, an exception is thrown. This is the user that is returned when the new user (with the linked identity provider attribute) signs in. The ProviderAttributeValue for the DestinationUser must match the username for the user in the user pool. The ProviderAttributeName will always be ignored.

alpfuSourceUser - An external identity provider account for a user who does not currently exist yet in the user pool. This user must be a federated user (for example, a SAML or Facebook user), not another native user. If the SourceUser is a federated social identity provider user (Facebook, Google, or Login with Amazon), you must set the ProviderAttributeName to Cognito_Subject . For social identity providers, the ProviderName will be Facebook , Google , or LoginWithAmazon , and Cognito will automatically parse the Facebook, Google, and Login with Amazon tokens for id , sub , and user_id , respectively. The ProviderAttributeValue for the user must be the same value as the id , sub , or user_id value found in the social identity provider token. For SAML, the ProviderAttributeName can be any value that matches a claim in the SAML assertion. If you wish to link SAML users based on the subject of the SAML assertion, you should map the subject to a claim through the SAML identity provider and submit that claim name as the ProviderAttributeName . If you set ProviderAttributeName to Cognito_Subject , Cognito will automatically parse the default unique identifier found in the subject from the SAML token.

data AdminLinkProviderForUser

See: adminLinkProviderForUser smart constructor.

Request Lenses

alpfuUserPoolId :: Lens' AdminLinkProviderForUser Text

The user pool ID for the user pool.

alpfuDestinationUser :: Lens' AdminLinkProviderForUser ProviderUserIdentifierType

The existing user in the user pool to be linked to the external identity provider user account. Can be a native (Username + Password) Cognito User Pools user or a federated user (for example, a SAML or Facebook user). If the user doesn't exist, an exception is thrown. This is the user that is returned when the new user (with the linked identity provider attribute) signs in. The ProviderAttributeValue for the DestinationUser must match the username for the user in the user pool. The ProviderAttributeName will always be ignored.

alpfuSourceUser :: Lens' AdminLinkProviderForUser ProviderUserIdentifierType

An external identity provider account for a user who does not currently exist yet in the user pool. This user must be a federated user (for example, a SAML or Facebook user), not another native user. If the SourceUser is a federated social identity provider user (Facebook, Google, or Login with Amazon), you must set the ProviderAttributeName to Cognito_Subject . For social identity providers, the ProviderName will be Facebook , Google , or LoginWithAmazon , and Cognito will automatically parse the Facebook, Google, and Login with Amazon tokens for id , sub , and user_id , respectively. The ProviderAttributeValue for the user must be the same value as the id , sub , or user_id value found in the social identity provider token. For SAML, the ProviderAttributeName can be any value that matches a claim in the SAML assertion. If you wish to link SAML users based on the subject of the SAML assertion, you should map the subject to a claim through the SAML identity provider and submit that claim name as the ProviderAttributeName . If you set ProviderAttributeName to Cognito_Subject , Cognito will automatically parse the default unique identifier found in the subject from the SAML token.

Destructuring the Response

adminLinkProviderForUserResponse

Arguments

:: Int
alpfursResponseStatus

-> AdminLinkProviderForUserResponse

Creates a value of AdminLinkProviderForUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

alpfursResponseStatus - -- | The response status code.

data AdminLinkProviderForUserResponse

See: adminLinkProviderForUserResponse smart constructor.

Response Lenses

alpfursResponseStatus :: Lens' AdminLinkProviderForUserResponse Int

- | The response status code.