Network.AWS.CognitoIdentityProvider.AdminListDevices

Description

Lists devices, as an administrator.

Requires developer credentials.

Creating a Request

adminListDevices

Arguments

:: Text
aldUserPoolId

-> Text
aldUsername

-> AdminListDevices

Creates a value of AdminListDevices with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aldPaginationToken - The pagination token.

aldLimit - The limit of the devices request.

aldUserPoolId - The user pool ID.

aldUsername - The user name.

data AdminListDevices

Represents the request to list devices, as an administrator.

See: adminListDevices smart constructor.

Request Lenses

aldPaginationToken :: Lens' AdminListDevices (Maybe Text)

The pagination token.

aldLimit :: Lens' AdminListDevices (Maybe Natural)

The limit of the devices request.

aldUserPoolId :: Lens' AdminListDevices Text

The user pool ID.

aldUsername :: Lens' AdminListDevices Text

The user name.

Destructuring the Response

adminListDevicesResponse

Arguments

:: Int
aldrsResponseStatus

-> AdminListDevicesResponse

Creates a value of AdminListDevicesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aldrsPaginationToken - The pagination token.

aldrsDevices - The devices in the list of devices response.

aldrsResponseStatus - -- | The response status code.

data AdminListDevicesResponse

Lists the device's response, as an administrator.

See: adminListDevicesResponse smart constructor.

Response Lenses

aldrsPaginationToken :: Lens' AdminListDevicesResponse (Maybe Text)

The pagination token.

aldrsDevices :: Lens' AdminListDevicesResponse [DeviceType]

The devices in the list of devices response.

aldrsResponseStatus :: Lens' AdminListDevicesResponse Int

- | The response status code.