Network.AWS.CognitoIdentityProvider.AdminListGroupsForUser

Description

Lists the groups that the user belongs to.

Requires developer credentials.

Creating a Request

adminListGroupsForUser

Arguments

:: Text
algfuUsername

-> Text
algfuUserPoolId

-> AdminListGroupsForUser

Creates a value of AdminListGroupsForUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

algfuNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

algfuLimit - The limit of the request to list groups.

algfuUsername - The username for the user.

algfuUserPoolId - The user pool ID for the user pool.

data AdminListGroupsForUser

See: adminListGroupsForUser smart constructor.

Request Lenses

algfuNextToken :: Lens' AdminListGroupsForUser (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

algfuLimit :: Lens' AdminListGroupsForUser (Maybe Natural)

The limit of the request to list groups.

algfuUsername :: Lens' AdminListGroupsForUser Text

The username for the user.

algfuUserPoolId :: Lens' AdminListGroupsForUser Text

The user pool ID for the user pool.

Destructuring the Response

adminListGroupsForUserResponse

Arguments

:: Int
algfursResponseStatus

-> AdminListGroupsForUserResponse

Creates a value of AdminListGroupsForUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

algfursGroups - The groups that the user belongs to.

algfursNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

algfursResponseStatus - -- | The response status code.

data AdminListGroupsForUserResponse

See: adminListGroupsForUserResponse smart constructor.

Response Lenses

algfursGroups :: Lens' AdminListGroupsForUserResponse [GroupType]

The groups that the user belongs to.

algfursNextToken :: Lens' AdminListGroupsForUserResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

algfursResponseStatus :: Lens' AdminListGroupsForUserResponse Int

- | The response status code.