Network.AWS.CognitoIdentityProvider.AdminRespondToAuthChallenge

Description

Responds to an authentication challenge, as an administrator.

Requires developer credentials.

Creating a Request

adminRespondToAuthChallenge

Arguments

:: Text
artacUserPoolId

-> Text
artacClientId

-> ChallengeNameType
artacChallengeName

-> AdminRespondToAuthChallenge

Creates a value of AdminRespondToAuthChallenge with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

artacChallengeResponses - The challenge responses. These are inputs corresponding to the value of ChallengeName , for example: * SMS_MFA : SMS_MFA_CODE , USERNAME , SECRET_HASH (if app client is configured with client secret). * PASSWORD_VERIFIER : PASSWORD_CLAIM_SIGNATURE , PASSWORD_CLAIM_SECRET_BLOCK , TIMESTAMP , USERNAME , SECRET_HASH (if app client is configured with client secret). * ADMIN_NO_SRP_AUTH : PASSWORD , USERNAME , SECRET_HASH (if app client is configured with client secret). * NEW_PASSWORD_REQUIRED : NEW_PASSWORD , any other required attributes, USERNAME , SECRET_HASH (if app client is configured with client secret). The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth .

artacSession - The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

artacUserPoolId - The ID of the Amazon Cognito user pool.

artacClientId - The app client ID.

artacChallengeName - The challenge name. For more information, see AdminInitiateAuth .

data AdminRespondToAuthChallenge

The request to respond to the authentication challenge, as an administrator.

See: adminRespondToAuthChallenge smart constructor.

Request Lenses

artacChallengeResponses :: Lens' AdminRespondToAuthChallenge (HashMap Text Text)

The challenge responses. These are inputs corresponding to the value of ChallengeName , for example: * SMS_MFA : SMS_MFA_CODE , USERNAME , SECRET_HASH (if app client is configured with client secret). * PASSWORD_VERIFIER : PASSWORD_CLAIM_SIGNATURE , PASSWORD_CLAIM_SECRET_BLOCK , TIMESTAMP , USERNAME , SECRET_HASH (if app client is configured with client secret). * ADMIN_NO_SRP_AUTH : PASSWORD , USERNAME , SECRET_HASH (if app client is configured with client secret). * NEW_PASSWORD_REQUIRED : NEW_PASSWORD , any other required attributes, USERNAME , SECRET_HASH (if app client is configured with client secret). The value of the USERNAME attribute must be the user's actual username, not an alias (such as email address or phone number). To make this easier, the AdminInitiateAuth response includes the actual username value in the USERNAMEUSER_ID_FOR_SRP attribute, even if you specified an alias in your call to AdminInitiateAuth .

artacSession :: Lens' AdminRespondToAuthChallenge (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

artacUserPoolId :: Lens' AdminRespondToAuthChallenge Text

The ID of the Amazon Cognito user pool.

artacClientId :: Lens' AdminRespondToAuthChallenge Text

The app client ID.

artacChallengeName :: Lens' AdminRespondToAuthChallenge ChallengeNameType

The challenge name. For more information, see AdminInitiateAuth .

Destructuring the Response

adminRespondToAuthChallengeResponse

Arguments

:: Int
artacrsResponseStatus

-> AdminRespondToAuthChallengeResponse

Creates a value of AdminRespondToAuthChallengeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

artacrsChallengeName - The name of the challenge. For more information, see AdminInitiateAuth .

artacrsChallengeParameters - The challenge parameters. For more information, see AdminInitiateAuth .

artacrsAuthenticationResult - The result returned by the server in response to the authentication request.

artacrsSession - The session which should be passed both ways in challenge-response calls to the service. If the InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

artacrsResponseStatus - -- | The response status code.

data AdminRespondToAuthChallengeResponse

Responds to the authentication challenge, as an administrator.

See: adminRespondToAuthChallengeResponse smart constructor.

Response Lenses

artacrsChallengeName :: Lens' AdminRespondToAuthChallengeResponse (Maybe ChallengeNameType)

The name of the challenge. For more information, see AdminInitiateAuth .

artacrsChallengeParameters :: Lens' AdminRespondToAuthChallengeResponse (HashMap Text Text)

The challenge parameters. For more information, see AdminInitiateAuth .

artacrsAuthenticationResult :: Lens' AdminRespondToAuthChallengeResponse (Maybe AuthenticationResultType)

The result returned by the server in response to the authentication request.

artacrsSession :: Lens' AdminRespondToAuthChallengeResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If the InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

artacrsResponseStatus :: Lens' AdminRespondToAuthChallengeResponse Int

- | The response status code.