Network.AWS.CognitoIdentityProvider.AdminUpdateDeviceStatus

Description

Updates the device status as an administrator.

Requires developer credentials.

Creating a Request

adminUpdateDeviceStatus

Arguments

:: Text
audsUserPoolId

-> Text
audsUsername

-> Text
audsDeviceKey

-> AdminUpdateDeviceStatus

Creates a value of AdminUpdateDeviceStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

audsDeviceRememberedStatus - The status indicating whether a device has been remembered or not.

audsUserPoolId - The user pool ID.

audsUsername - The user name.

audsDeviceKey - The device key.

data AdminUpdateDeviceStatus

The request to update the device status, as an administrator.

See: adminUpdateDeviceStatus smart constructor.

Request Lenses

audsDeviceRememberedStatus :: Lens' AdminUpdateDeviceStatus (Maybe DeviceRememberedStatusType)

The status indicating whether a device has been remembered or not.

audsUserPoolId :: Lens' AdminUpdateDeviceStatus Text

The user pool ID.

audsUsername :: Lens' AdminUpdateDeviceStatus Text

The user name.

audsDeviceKey :: Lens' AdminUpdateDeviceStatus Text

The device key.

Destructuring the Response

adminUpdateDeviceStatusResponse

Arguments

:: Int
audsrsResponseStatus

-> AdminUpdateDeviceStatusResponse

Creates a value of AdminUpdateDeviceStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

audsrsResponseStatus - -- | The response status code.

data AdminUpdateDeviceStatusResponse

The status response from the request to update the device, as an administrator.

See: adminUpdateDeviceStatusResponse smart constructor.

Response Lenses

audsrsResponseStatus :: Lens' AdminUpdateDeviceStatusResponse Int

- | The response status code.