Network.AWS.CognitoIdentityProvider.AdminUserGlobalSignOut

Description

Signs out users from all devices, as an administrator.

Requires developer credentials.

Creating a Request

adminUserGlobalSignOut

Arguments

:: Text
augsoUserPoolId

-> Text
augsoUsername

-> AdminUserGlobalSignOut

Creates a value of AdminUserGlobalSignOut with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

augsoUserPoolId - The user pool ID.

augsoUsername - The user name.

data AdminUserGlobalSignOut

The request to sign out of all devices, as an administrator.

See: adminUserGlobalSignOut smart constructor.

Request Lenses

augsoUserPoolId :: Lens' AdminUserGlobalSignOut Text

The user pool ID.

augsoUsername :: Lens' AdminUserGlobalSignOut Text

The user name.

Destructuring the Response

adminUserGlobalSignOutResponse

Arguments

:: Int
augsorsResponseStatus

-> AdminUserGlobalSignOutResponse

Creates a value of AdminUserGlobalSignOutResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

augsorsResponseStatus - -- | The response status code.

data AdminUserGlobalSignOutResponse

The global sign-out response, as an administrator.

See: adminUserGlobalSignOutResponse smart constructor.

Response Lenses

augsorsResponseStatus :: Lens' AdminUserGlobalSignOutResponse Int

- | The response status code.