Network.AWS.CognitoIdentityProvider.ChangePassword

Description

Changes the password for a specified user in a user pool.

Creating a Request

changePassword

Arguments

:: Text
cpPreviousPassword

-> Text
cpProposedPassword

-> Text
cpAccessToken

-> ChangePassword

Creates a value of ChangePassword with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpPreviousPassword - The old password in the change password request.

cpProposedPassword - The new password in the change password request.

cpAccessToken - The access token in the change password request.

data ChangePassword

Represents the request to change a user password.

See: changePassword smart constructor.

Request Lenses

cpPreviousPassword :: Lens' ChangePassword Text

The old password in the change password request.

cpProposedPassword :: Lens' ChangePassword Text

The new password in the change password request.

cpAccessToken :: Lens' ChangePassword Text

The access token in the change password request.

Destructuring the Response

changePasswordResponse

Arguments

:: Int
cprsResponseStatus

-> ChangePasswordResponse

Creates a value of ChangePasswordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cprsResponseStatus - -- | The response status code.

data ChangePasswordResponse

The response from the server to the change password request.

See: changePasswordResponse smart constructor.

Response Lenses

cprsResponseStatus :: Lens' ChangePasswordResponse Int

- | The response status code.