Network.AWS.CognitoIdentityProvider.ConfirmForgotPassword

Description

Allows a user to enter a confirmation code to reset a forgotten password.

Creating a Request

confirmForgotPassword

Arguments

:: Text
cfpClientId

-> Text
cfpUsername

-> Text
cfpConfirmationCode

-> Text
cfpPassword

-> ConfirmForgotPassword

Creates a value of ConfirmForgotPassword with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfpSecretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

cfpClientId - The app client ID of the app associated with the user pool.

cfpUsername - The user name of the user for whom you want to enter a code to retrieve a forgotten password.

cfpConfirmationCode - The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see ForgotPassword

cfpPassword - The password sent by a user's request to retrieve a forgotten password.

data ConfirmForgotPassword

The request representing the confirmation for a password reset.

See: confirmForgotPassword smart constructor.

Request Lenses

cfpSecretHash :: Lens' ConfirmForgotPassword (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

cfpClientId :: Lens' ConfirmForgotPassword Text

The app client ID of the app associated with the user pool.

cfpUsername :: Lens' ConfirmForgotPassword Text

The user name of the user for whom you want to enter a code to retrieve a forgotten password.

cfpConfirmationCode :: Lens' ConfirmForgotPassword Text

The confirmation code sent by a user's request to retrieve a forgotten password. For more information, see ForgotPassword

cfpPassword :: Lens' ConfirmForgotPassword Text

The password sent by a user's request to retrieve a forgotten password.

Destructuring the Response

confirmForgotPasswordResponse

Arguments

:: Int
cfprsResponseStatus

-> ConfirmForgotPasswordResponse

Creates a value of ConfirmForgotPasswordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfprsResponseStatus - -- | The response status code.

data ConfirmForgotPasswordResponse

The response from the server that results from a user's request to retrieve a forgotten password.

See: confirmForgotPasswordResponse smart constructor.

Response Lenses

cfprsResponseStatus :: Lens' ConfirmForgotPasswordResponse Int

- | The response status code.