Network.AWS.CognitoIdentityProvider.ConfirmSignUp

Description

Confirms registration of a user and handles the existing alias from a previous user.

Creating a Request

confirmSignUp

Arguments

:: Text
csuClientId

-> Text
csuUsername

-> Text
csuConfirmationCode

-> ConfirmSignUp

Creates a value of ConfirmSignUp with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csuForceAliasCreation - Boolean to be specified to force user confirmation irrespective of existing alias. By default set to False . If this parameter is set to True and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to False , the API will throw an AliasExistsException error.

csuSecretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

csuClientId - The ID of the app client associated with the user pool.

csuUsername - The user name of the user whose registration you wish to confirm.

csuConfirmationCode - The confirmation code sent by a user's request to confirm registration.

data ConfirmSignUp

Represents the request to confirm registration of a user.

See: confirmSignUp smart constructor.

Request Lenses

csuForceAliasCreation :: Lens' ConfirmSignUp (Maybe Bool)

Boolean to be specified to force user confirmation irrespective of existing alias. By default set to False . If this parameter is set to True and the phone number/email used for sign up confirmation already exists as an alias with a different user, the API call will migrate the alias from the previous user to the newly created user being confirmed. If set to False , the API will throw an AliasExistsException error.

csuSecretHash :: Lens' ConfirmSignUp (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

csuClientId :: Lens' ConfirmSignUp Text

The ID of the app client associated with the user pool.

csuUsername :: Lens' ConfirmSignUp Text

The user name of the user whose registration you wish to confirm.

csuConfirmationCode :: Lens' ConfirmSignUp Text

The confirmation code sent by a user's request to confirm registration.

Destructuring the Response

confirmSignUpResponse

Arguments

:: Int
csursResponseStatus

-> ConfirmSignUpResponse

Creates a value of ConfirmSignUpResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csursResponseStatus - -- | The response status code.

data ConfirmSignUpResponse

Represents the response from the server for the registration confirmation.

See: confirmSignUpResponse smart constructor.

Response Lenses

csursResponseStatus :: Lens' ConfirmSignUpResponse Int

- | The response status code.