Network.AWS.CognitoIdentityProvider.CreateGroup

Description

Creates a new group in the specified user pool.

Requires developer credentials.

Creating a Request

createGroup

Arguments

:: Text
cgGroupName

-> Text
cgUserPoolId

-> CreateGroup

Creates a value of CreateGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgPrecedence - A nonnegative integer value that specifies the precedence of this group relative to the other groups that a user can belong to in the user pool. Zero is the highest precedence value. Groups with lower Precedence values take precedence over groups with higher or null Precedence values. If a user belongs to two or more groups, it is the group with the lowest precedence value whose role ARN will be used in the cognito:roles and cognito:preferred_role claims in the user's tokens. Two groups can have the same Precedence value. If this happens, neither group takes precedence over the other. If two groups with the same Precedence have the same role ARN, that role is used in the cognito:preferred_role claim in tokens for users in each group. If the two groups have different role ARNs, the cognito:preferred_role claim is not set in users' tokens. The default Precedence value is null.

cgDescription - A string containing the description of the group.

cgRoleARN - The role ARN for the group.

cgGroupName - The name of the group. Must be unique.

cgUserPoolId - The user pool ID for the user pool.

data CreateGroup

See: createGroup smart constructor.

Request Lenses

cgPrecedence :: Lens' CreateGroup (Maybe Natural)

A nonnegative integer value that specifies the precedence of this group relative to the other groups that a user can belong to in the user pool. Zero is the highest precedence value. Groups with lower Precedence values take precedence over groups with higher or null Precedence values. If a user belongs to two or more groups, it is the group with the lowest precedence value whose role ARN will be used in the cognito:roles and cognito:preferred_role claims in the user's tokens. Two groups can have the same Precedence value. If this happens, neither group takes precedence over the other. If two groups with the same Precedence have the same role ARN, that role is used in the cognito:preferred_role claim in tokens for users in each group. If the two groups have different role ARNs, the cognito:preferred_role claim is not set in users' tokens. The default Precedence value is null.

cgDescription :: Lens' CreateGroup (Maybe Text)

A string containing the description of the group.

cgRoleARN :: Lens' CreateGroup (Maybe Text)

The role ARN for the group.

cgGroupName :: Lens' CreateGroup Text

The name of the group. Must be unique.

cgUserPoolId :: Lens' CreateGroup Text

The user pool ID for the user pool.

Destructuring the Response

createGroupResponse

Arguments

:: Int
cgrsResponseStatus

-> CreateGroupResponse

Creates a value of CreateGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgrsGroup - The group object for the group.

cgrsResponseStatus - -- | The response status code.

data CreateGroupResponse

See: createGroupResponse smart constructor.

Response Lenses

cgrsGroup :: Lens' CreateGroupResponse (Maybe GroupType)

The group object for the group.

cgrsResponseStatus :: Lens' CreateGroupResponse Int

- | The response status code.