Network.AWS.CognitoIdentityProvider.CreateIdentityProvider

Description

Creates an identity provider for a user pool.

Creating a Request

createIdentityProvider

Arguments

:: Text
cipUserPoolId

-> Text
cipProviderName

-> IdentityProviderTypeType
cipProviderType

-> CreateIdentityProvider

Creates a value of CreateIdentityProvider with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cipIdpIdentifiers - A list of identity provider identifiers.

cipAttributeMapping - A mapping of identity provider attributes to standard and custom user pool attributes.

cipUserPoolId - The user pool ID.

cipProviderName - The identity provider name.

cipProviderType - The identity provider type.

cipProviderDetails - The identity provider details, such as MetadataURL and MetadataFile .

data CreateIdentityProvider

See: createIdentityProvider smart constructor.

Request Lenses

cipIdpIdentifiers :: Lens' CreateIdentityProvider [Text]

A list of identity provider identifiers.

cipAttributeMapping :: Lens' CreateIdentityProvider (HashMap Text Text)

A mapping of identity provider attributes to standard and custom user pool attributes.

cipUserPoolId :: Lens' CreateIdentityProvider Text

The user pool ID.

cipProviderName :: Lens' CreateIdentityProvider Text

The identity provider name.

cipProviderType :: Lens' CreateIdentityProvider IdentityProviderTypeType

The identity provider type.

cipProviderDetails :: Lens' CreateIdentityProvider (HashMap Text Text)

The identity provider details, such as MetadataURL and MetadataFile .

Destructuring the Response

createIdentityProviderResponse

Arguments

:: Int
ciprsResponseStatus

-> IdentityProviderType
ciprsIdentityProvider

-> CreateIdentityProviderResponse

Creates a value of CreateIdentityProviderResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciprsResponseStatus - -- | The response status code.

ciprsIdentityProvider - The newly created identity provider object.

data CreateIdentityProviderResponse

See: createIdentityProviderResponse smart constructor.

Response Lenses

ciprsResponseStatus :: Lens' CreateIdentityProviderResponse Int

- | The response status code.

ciprsIdentityProvider :: Lens' CreateIdentityProviderResponse IdentityProviderType

The newly created identity provider object.