Network.AWS.CognitoIdentityProvider.CreateUserPool

Description

Creates a new Amazon Cognito user pool and sets the password policy for the pool.

Creating a Request

createUserPool

Arguments

:: Text
cupPoolName

-> CreateUserPool

Creates a value of CreateUserPool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupUserPoolTags - The cost allocation tags for the user pool. For more information, see Adding Cost Allocation Tags to Your User Pool

cupVerificationMessageTemplate - The template for the verification message that the user sees when the app requests permission to access the user's information.

cupEmailVerificationMessage - A string representing the email verification message.

cupSmsAuthenticationMessage - A string representing the SMS authentication message.

cupEmailVerificationSubject - A string representing the email verification subject.

cupUsernameAttributes - Specifies whether email addresses or phone numbers can be specified as usernames when a user signs up.

cupAliasAttributes - Attributes supported as an alias for this user pool. Possible values: phone_number , email , or preferred_username .

cupSchema - An array of schema attributes for the new user pool. These attributes can be standard or custom attributes.

cupEmailConfiguration - The email configuration.

cupSmsVerificationMessage - A string representing the SMS verification message.

cupMFAConfiguration - Specifies MFA configuration details.

cupLambdaConfig - The Lambda trigger configuration information for the new user pool.

cupSmsConfiguration - The SMS configuration.

cupAdminCreateUserConfig - The configuration for AdminCreateUser requests.

cupDeviceConfiguration - The device configuration.

cupAutoVerifiedAttributes - The attributes to be auto-verified. Possible values: email , phone_number .

cupPolicies - The policies associated with the new user pool.

cupPoolName - A string used to name the user pool.

data CreateUserPool

Represents the request to create a user pool.

See: createUserPool smart constructor.

Request Lenses

cupUserPoolTags :: Lens' CreateUserPool (HashMap Text Text)

The cost allocation tags for the user pool. For more information, see Adding Cost Allocation Tags to Your User Pool

cupVerificationMessageTemplate :: Lens' CreateUserPool (Maybe VerificationMessageTemplateType)

The template for the verification message that the user sees when the app requests permission to access the user's information.

cupEmailVerificationMessage :: Lens' CreateUserPool (Maybe Text)

A string representing the email verification message.

cupSmsAuthenticationMessage :: Lens' CreateUserPool (Maybe Text)

A string representing the SMS authentication message.

cupEmailVerificationSubject :: Lens' CreateUserPool (Maybe Text)

A string representing the email verification subject.

cupUsernameAttributes :: Lens' CreateUserPool [UsernameAttributeType]

Specifies whether email addresses or phone numbers can be specified as usernames when a user signs up.

cupAliasAttributes :: Lens' CreateUserPool [AliasAttributeType]

Attributes supported as an alias for this user pool. Possible values: phone_number , email , or preferred_username .

cupSchema :: Lens' CreateUserPool (Maybe (NonEmpty SchemaAttributeType))

An array of schema attributes for the new user pool. These attributes can be standard or custom attributes.

cupEmailConfiguration :: Lens' CreateUserPool (Maybe EmailConfigurationType)

The email configuration.

cupSmsVerificationMessage :: Lens' CreateUserPool (Maybe Text)

A string representing the SMS verification message.

cupMFAConfiguration :: Lens' CreateUserPool (Maybe UserPoolMFAType)

Specifies MFA configuration details.

cupLambdaConfig :: Lens' CreateUserPool (Maybe LambdaConfigType)

The Lambda trigger configuration information for the new user pool.

cupSmsConfiguration :: Lens' CreateUserPool (Maybe SmsConfigurationType)

The SMS configuration.

cupAdminCreateUserConfig :: Lens' CreateUserPool (Maybe AdminCreateUserConfigType)

The configuration for AdminCreateUser requests.

cupDeviceConfiguration :: Lens' CreateUserPool (Maybe DeviceConfigurationType)

The device configuration.

cupAutoVerifiedAttributes :: Lens' CreateUserPool [VerifiedAttributeType]

The attributes to be auto-verified. Possible values: email , phone_number .

cupPolicies :: Lens' CreateUserPool (Maybe UserPoolPolicyType)

The policies associated with the new user pool.

cupPoolName :: Lens' CreateUserPool Text

A string used to name the user pool.

Destructuring the Response

createUserPoolResponse

Arguments

:: Int
cuprsResponseStatus

-> CreateUserPoolResponse

Creates a value of CreateUserPoolResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cuprsUserPool - A container for the user pool details.

cuprsResponseStatus - -- | The response status code.

data CreateUserPoolResponse

Represents the response from the server for the request to create a user pool.

See: createUserPoolResponse smart constructor.

Response Lenses

cuprsUserPool :: Lens' CreateUserPoolResponse (Maybe UserPoolType)

A container for the user pool details.

cuprsResponseStatus :: Lens' CreateUserPoolResponse Int

- | The response status code.