Network.AWS.CognitoIdentityProvider.CreateUserPoolClient

Description

Creates the user pool client.

Creating a Request

createUserPoolClient

Arguments

:: Text
cupcUserPoolId

-> Text
cupcClientName

-> CreateUserPoolClient

Creates a value of CreateUserPoolClient with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupcRefreshTokenValidity - The time limit, in days, after which the refresh token is no longer valid and cannot be used.

cupcExplicitAuthFlows - The explicit authentication flows.

cupcSupportedIdentityProviders - A list of provider names for the identity providers that are supported on this client.

cupcLogoutURLs - A list of allowed logout URLs for the identity providers.

cupcAllowedOAuthFlowsUserPoolClient - Set to True if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

cupcGenerateSecret - Boolean to specify whether you want to generate a secret for the user pool client being created.

cupcDefaultRedirectURI - The default redirect URI. Must be in the CallbackURLs list.

cupcWriteAttributes - The write attributes.

cupcReadAttributes - The read attributes.

cupcAllowedOAuthScopes - A list of allowed OAuth scopes. Currently supported values are "phone" , "email" , "openid" , and Cognito .

cupcAllowedOAuthFlows - Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint. Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

cupcCallbackURLs - A list of allowed callback URLs for the identity providers.

cupcUserPoolId - The user pool ID for the user pool where you want to create a user pool client.

cupcClientName - The client name for the user pool client you would like to create.

data CreateUserPoolClient

Represents the request to create a user pool client.

See: createUserPoolClient smart constructor.

Request Lenses

cupcRefreshTokenValidity :: Lens' CreateUserPoolClient (Maybe Natural)

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

cupcExplicitAuthFlows :: Lens' CreateUserPoolClient [ExplicitAuthFlowsType]

The explicit authentication flows.

cupcSupportedIdentityProviders :: Lens' CreateUserPoolClient [Text]

A list of provider names for the identity providers that are supported on this client.

cupcLogoutURLs :: Lens' CreateUserPoolClient [Text]

A list of allowed logout URLs for the identity providers.

cupcAllowedOAuthFlowsUserPoolClient :: Lens' CreateUserPoolClient (Maybe Bool)

Set to True if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

cupcGenerateSecret :: Lens' CreateUserPoolClient (Maybe Bool)

Boolean to specify whether you want to generate a secret for the user pool client being created.

cupcDefaultRedirectURI :: Lens' CreateUserPoolClient (Maybe Text)

The default redirect URI. Must be in the CallbackURLs list.

cupcWriteAttributes :: Lens' CreateUserPoolClient [Text]

The write attributes.

cupcReadAttributes :: Lens' CreateUserPoolClient [Text]

The read attributes.

cupcAllowedOAuthScopes :: Lens' CreateUserPoolClient [Text]

A list of allowed OAuth scopes. Currently supported values are "phone" , "email" , "openid" , and Cognito .

cupcAllowedOAuthFlows :: Lens' CreateUserPoolClient [OAuthFlowType]

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint. Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

cupcCallbackURLs :: Lens' CreateUserPoolClient [Text]

A list of allowed callback URLs for the identity providers.

cupcUserPoolId :: Lens' CreateUserPoolClient Text

The user pool ID for the user pool where you want to create a user pool client.

cupcClientName :: Lens' CreateUserPoolClient Text

The client name for the user pool client you would like to create.

Destructuring the Response

createUserPoolClientResponse

Arguments

:: Int
cupcrsResponseStatus

-> CreateUserPoolClientResponse

Creates a value of CreateUserPoolClientResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupcrsUserPoolClient - The user pool client that was just created.

cupcrsResponseStatus - -- | The response status code.

data CreateUserPoolClientResponse

Represents the response from the server to create a user pool client.

See: createUserPoolClientResponse smart constructor.

Response Lenses

cupcrsUserPoolClient :: Lens' CreateUserPoolClientResponse (Maybe UserPoolClientType)

The user pool client that was just created.

cupcrsResponseStatus :: Lens' CreateUserPoolClientResponse Int

- | The response status code.