Network.AWS.CognitoIdentityProvider.CreateUserPoolDomain

Description

Creates a new domain for a user pool.

Creating a Request

createUserPoolDomain

Arguments

:: Text
cupdDomain

-> Text
cupdUserPoolId

-> CreateUserPoolDomain

Creates a value of CreateUserPoolDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupdDomain - The domain string.

cupdUserPoolId - The user pool ID.

data CreateUserPoolDomain

See: createUserPoolDomain smart constructor.

Request Lenses

cupdDomain :: Lens' CreateUserPoolDomain Text

The domain string.

cupdUserPoolId :: Lens' CreateUserPoolDomain Text

The user pool ID.

Destructuring the Response

createUserPoolDomainResponse

Arguments

:: Int
cupdrsResponseStatus

-> CreateUserPoolDomainResponse

Creates a value of CreateUserPoolDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cupdrsResponseStatus - -- | The response status code.

data CreateUserPoolDomainResponse

See: createUserPoolDomainResponse smart constructor.

Response Lenses

cupdrsResponseStatus :: Lens' CreateUserPoolDomainResponse Int

- | The response status code.