Network.AWS.CognitoIdentityProvider.DeleteResourceServer

Description

Deletes a resource server.

Creating a Request

deleteResourceServer

Arguments

:: Text
drsUserPoolId

-> Text
drsIdentifier

-> DeleteResourceServer

Creates a value of DeleteResourceServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsUserPoolId - The user pool ID for the user pool that hosts the resource server.

drsIdentifier - The identifier for the resource server.

data DeleteResourceServer

See: deleteResourceServer smart constructor.

Request Lenses

drsUserPoolId :: Lens' DeleteResourceServer Text

The user pool ID for the user pool that hosts the resource server.

drsIdentifier :: Lens' DeleteResourceServer Text

The identifier for the resource server.

Destructuring the Response

deleteResourceServerResponse :: DeleteResourceServerResponse

Creates a value of DeleteResourceServerResponse with the minimum fields required to make a request.

data DeleteResourceServerResponse

See: deleteResourceServerResponse smart constructor.