Network.AWS.CognitoIdentityProvider.DeleteUserAttributes

Description

Deletes the attributes for a user.

Creating a Request

deleteUserAttributes

Arguments

:: Text
duaAccessToken

-> DeleteUserAttributes

Creates a value of DeleteUserAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

duaUserAttributeNames - An array of strings representing the user attribute names you wish to delete. For custom attributes, you must prepend the custom: prefix to the attribute name.

duaAccessToken - The access token used in the request to delete user attributes.

data DeleteUserAttributes

Represents the request to delete user attributes.

See: deleteUserAttributes smart constructor.

Request Lenses

duaUserAttributeNames :: Lens' DeleteUserAttributes [Text]

An array of strings representing the user attribute names you wish to delete. For custom attributes, you must prepend the custom: prefix to the attribute name.

duaAccessToken :: Lens' DeleteUserAttributes Text

The access token used in the request to delete user attributes.

Destructuring the Response

deleteUserAttributesResponse

Arguments

:: Int
duarsResponseStatus

-> DeleteUserAttributesResponse

Creates a value of DeleteUserAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

duarsResponseStatus - -- | The response status code.

data DeleteUserAttributesResponse

Represents the response from the server to delete user attributes.

See: deleteUserAttributesResponse smart constructor.

Response Lenses

duarsResponseStatus :: Lens' DeleteUserAttributesResponse Int

- | The response status code.