Network.AWS.CognitoIdentityProvider.DeleteUserPoolDomain

Description

Deletes a domain for a user pool.

Creating a Request

deleteUserPoolDomain

Arguments

:: Text
dupdDomain

-> Text
dupdUserPoolId

-> DeleteUserPoolDomain

Creates a value of DeleteUserPoolDomain with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dupdDomain - The domain string.

dupdUserPoolId - The user pool ID.

data DeleteUserPoolDomain

See: deleteUserPoolDomain smart constructor.

Request Lenses

dupdDomain :: Lens' DeleteUserPoolDomain Text

The domain string.

dupdUserPoolId :: Lens' DeleteUserPoolDomain Text

The user pool ID.

Destructuring the Response

deleteUserPoolDomainResponse

Arguments

:: Int
dupdrsResponseStatus

-> DeleteUserPoolDomainResponse

Creates a value of DeleteUserPoolDomainResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dupdrsResponseStatus - -- | The response status code.

data DeleteUserPoolDomainResponse

See: deleteUserPoolDomainResponse smart constructor.

Response Lenses

dupdrsResponseStatus :: Lens' DeleteUserPoolDomainResponse Int

- | The response status code.