Network.AWS.CognitoIdentityProvider.DescribeResourceServer

Description

Describes a resource server.

Creating a Request

describeResourceServer

Arguments

:: Text
desUserPoolId

-> Text
desIdentifier

-> DescribeResourceServer

Creates a value of DescribeResourceServer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desUserPoolId - The user pool ID for the user pool that hosts the resource server.

desIdentifier - The identifier for the resource server

data DescribeResourceServer

See: describeResourceServer smart constructor.

Request Lenses

desUserPoolId :: Lens' DescribeResourceServer Text

The user pool ID for the user pool that hosts the resource server.

desIdentifier :: Lens' DescribeResourceServer Text

The identifier for the resource server

Destructuring the Response

describeResourceServerResponse

Arguments

:: Int
drsrsResponseStatus

-> ResourceServerType
drsrsResourceServer

-> DescribeResourceServerResponse

Creates a value of DescribeResourceServerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsrsResponseStatus - -- | The response status code.

drsrsResourceServer - The resource server.

data DescribeResourceServerResponse

See: describeResourceServerResponse smart constructor.

Response Lenses

drsrsResponseStatus :: Lens' DescribeResourceServerResponse Int

- | The response status code.

drsrsResourceServer :: Lens' DescribeResourceServerResponse ResourceServerType

The resource server.