Network.AWS.CognitoIdentityProvider.ForgotPassword

Description

Calling this API causes a message to be sent to the end user with a confirmation code that is required to change the user's password. For the Username parameter, you can use the username or user alias. If a verified phone number exists for the user, the confirmation code is sent to the phone number. Otherwise, if a verified email exists, the confirmation code is sent to the email. If neither a verified phone number nor a verified email exists, InvalidParameterException is thrown. To use the confirmation code for resetting the password, call ConfirmForgotPassword .

Creating a Request

forgotPassword

Arguments

:: Text
fpClientId

-> Text
fpUsername

-> ForgotPassword

Creates a value of ForgotPassword with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fpSecretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

fpClientId - The ID of the client associated with the user pool.

fpUsername - The user name of the user for whom you want to enter a code to reset a forgotten password.

data ForgotPassword

Represents the request to reset a user's password.

See: forgotPassword smart constructor.

Request Lenses

fpSecretHash :: Lens' ForgotPassword (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

fpClientId :: Lens' ForgotPassword Text

The ID of the client associated with the user pool.

fpUsername :: Lens' ForgotPassword Text

The user name of the user for whom you want to enter a code to reset a forgotten password.

Destructuring the Response

forgotPasswordResponse

Arguments

:: Int
fprsResponseStatus

-> ForgotPasswordResponse

Creates a value of ForgotPasswordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fprsCodeDeliveryDetails - The code delivery details returned by the server in response to the request to reset a password.

fprsResponseStatus - -- | The response status code.

data ForgotPasswordResponse

Respresents the response from the server regarding the request to reset a password.

See: forgotPasswordResponse smart constructor.

Response Lenses

fprsCodeDeliveryDetails :: Lens' ForgotPasswordResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to reset a password.

fprsResponseStatus :: Lens' ForgotPasswordResponse Int

- | The response status code.