Network.AWS.CognitoIdentityProvider.GetDevice

Description

Gets the device.

Creating a Request

getDevice

Arguments

:: Text
gdDeviceKey

-> GetDevice

Creates a value of GetDevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdAccessToken - The access token.

gdDeviceKey - The device key.

data GetDevice

Represents the request to get the device.

See: getDevice smart constructor.

Request Lenses

gdAccessToken :: Lens' GetDevice (Maybe Text)

The access token.

gdDeviceKey :: Lens' GetDevice Text

The device key.

Destructuring the Response

getDeviceResponse

Arguments

:: Int
gdrsResponseStatus

-> DeviceType
gdrsDevice

-> GetDeviceResponse

Creates a value of GetDeviceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsResponseStatus - -- | The response status code.

gdrsDevice - The device.

data GetDeviceResponse

Gets the device response.

See: getDeviceResponse smart constructor.

Response Lenses

gdrsResponseStatus :: Lens' GetDeviceResponse Int

- | The response status code.

gdrsDevice :: Lens' GetDeviceResponse DeviceType

The device.