Network.AWS.CognitoIdentityProvider.GetIdentityProviderByIdentifier

Description

Gets the specified identity provider.

Creating a Request

getIdentityProviderByIdentifier

Arguments

:: Text
gipbiUserPoolId

-> Text
gipbiIdpIdentifier

-> GetIdentityProviderByIdentifier

Creates a value of GetIdentityProviderByIdentifier with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gipbiUserPoolId - The user pool ID.

gipbiIdpIdentifier - The identity provider ID.

data GetIdentityProviderByIdentifier

See: getIdentityProviderByIdentifier smart constructor.

Request Lenses

gipbiUserPoolId :: Lens' GetIdentityProviderByIdentifier Text

The user pool ID.

gipbiIdpIdentifier :: Lens' GetIdentityProviderByIdentifier Text

The identity provider ID.

Destructuring the Response

getIdentityProviderByIdentifierResponse

Arguments

:: Int
gipbirsResponseStatus

-> IdentityProviderType
gipbirsIdentityProvider

-> GetIdentityProviderByIdentifierResponse

Creates a value of GetIdentityProviderByIdentifierResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gipbirsResponseStatus - -- | The response status code.

gipbirsIdentityProvider - The identity provider object.

data GetIdentityProviderByIdentifierResponse

See: getIdentityProviderByIdentifierResponse smart constructor.

Response Lenses

gipbirsResponseStatus :: Lens' GetIdentityProviderByIdentifierResponse Int

- | The response status code.

gipbirsIdentityProvider :: Lens' GetIdentityProviderByIdentifierResponse IdentityProviderType

The identity provider object.