Network.AWS.CognitoIdentityProvider.GetUser

Description

Gets the user attributes and metadata for a user.

Creating a Request

getUser

Arguments

:: Text
guAccessToken

-> GetUser

Creates a value of GetUser with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guAccessToken - The access token returned by the server response to get information about the user.

data GetUser

Represents the request to get information about the user.

See: getUser smart constructor.

Request Lenses

guAccessToken :: Lens' GetUser Text

The access token returned by the server response to get information about the user.

Destructuring the Response

getUserResponse

Arguments

:: Int
gursResponseStatus

-> Text
gursUsername

-> GetUserResponse

Creates a value of GetUserResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gursMFAOptions - Specifies the options for MFA (e.g., email or phone number).

gursResponseStatus - -- | The response status code.

gursUsername - The user name of the user you wish to retrieve from the get user request.

gursUserAttributes - An array of name-value pairs representing user attributes. For custom attributes, you must prepend the custom: prefix to the attribute name.

data GetUserResponse

Represents the response from the server from the request to get information about the user.

See: getUserResponse smart constructor.

Response Lenses

gursMFAOptions :: Lens' GetUserResponse [MFAOptionType]

Specifies the options for MFA (e.g., email or phone number).

gursResponseStatus :: Lens' GetUserResponse Int

- | The response status code.

gursUsername :: Lens' GetUserResponse Text

The user name of the user you wish to retrieve from the get user request.

gursUserAttributes :: Lens' GetUserResponse [AttributeType]

An array of name-value pairs representing user attributes. For custom attributes, you must prepend the custom: prefix to the attribute name.