Network.AWS.CognitoIdentityProvider.GetUserAttributeVerificationCode

Description

Gets the user attribute verification code for the specified attribute name.

Creating a Request

getUserAttributeVerificationCode

Arguments

:: Text
guavcAccessToken

-> Text
guavcAttributeName

-> GetUserAttributeVerificationCode

Creates a value of GetUserAttributeVerificationCode with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guavcAccessToken - The access token returned by the server response to get the user attribute verification code.

guavcAttributeName - The attribute name returned by the server response to get the user attribute verification code.

data GetUserAttributeVerificationCode

Represents the request to get user attribute verification.

See: getUserAttributeVerificationCode smart constructor.

Request Lenses

guavcAccessToken :: Lens' GetUserAttributeVerificationCode Text

The access token returned by the server response to get the user attribute verification code.

guavcAttributeName :: Lens' GetUserAttributeVerificationCode Text

The attribute name returned by the server response to get the user attribute verification code.

Destructuring the Response

getUserAttributeVerificationCodeResponse

Arguments

:: Int
guavcrsResponseStatus

-> GetUserAttributeVerificationCodeResponse

Creates a value of GetUserAttributeVerificationCodeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

guavcrsCodeDeliveryDetails - The code delivery details returned by the server in response to the request to get the user attribute verification code.

guavcrsResponseStatus - -- | The response status code.

data GetUserAttributeVerificationCodeResponse

The verification code response returned by the server response to get the user attribute verification code.

See: getUserAttributeVerificationCodeResponse smart constructor.

Response Lenses

guavcrsCodeDeliveryDetails :: Lens' GetUserAttributeVerificationCodeResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to get the user attribute verification code.

guavcrsResponseStatus :: Lens' GetUserAttributeVerificationCodeResponse Int

- | The response status code.