Network.AWS.CognitoIdentityProvider.ListResourceServers

Description

Lists the resource servers for a user pool.

Creating a Request

listResourceServers

Arguments

:: Text
lrsUserPoolId

-> ListResourceServers

Creates a value of ListResourceServers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsNextToken - A pagination token.

lrsMaxResults - The maximum number of resource servers to return.

lrsUserPoolId - The user pool ID for the user pool.

data ListResourceServers

See: listResourceServers smart constructor.

Request Lenses

lrsNextToken :: Lens' ListResourceServers (Maybe Text)

A pagination token.

lrsMaxResults :: Lens' ListResourceServers (Maybe Natural)

The maximum number of resource servers to return.

lrsUserPoolId :: Lens' ListResourceServers Text

The user pool ID for the user pool.

Destructuring the Response

listResourceServersResponse

Arguments

:: Int
lrsrsResponseStatus

-> ListResourceServersResponse

Creates a value of ListResourceServersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsrsNextToken - A pagination token.

lrsrsResponseStatus - -- | The response status code.

lrsrsResourceServers - The resource servers.

data ListResourceServersResponse

See: listResourceServersResponse smart constructor.

Response Lenses

lrsrsNextToken :: Lens' ListResourceServersResponse (Maybe Text)

A pagination token.

lrsrsResponseStatus :: Lens' ListResourceServersResponse Int

- | The response status code.

lrsrsResourceServers :: Lens' ListResourceServersResponse [ResourceServerType]

The resource servers.