Network.AWS.CognitoIdentityProvider.ListUserImportJobs

Description

Lists the user import jobs.

Creating a Request

listUserImportJobs

Arguments

:: Text
luijUserPoolId

-> Natural
luijMaxResults

-> ListUserImportJobs

Creates a value of ListUserImportJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luijPaginationToken - An identifier that was returned from the previous call to ListUserImportJobs , which can be used to return the next set of import jobs in the list.

luijUserPoolId - The user pool ID for the user pool that the users are being imported into.

luijMaxResults - The maximum number of import jobs you want the request to return.

data ListUserImportJobs

Represents the request to list the user import jobs.

See: listUserImportJobs smart constructor.

Request Lenses

luijPaginationToken :: Lens' ListUserImportJobs (Maybe Text)

An identifier that was returned from the previous call to ListUserImportJobs , which can be used to return the next set of import jobs in the list.

luijUserPoolId :: Lens' ListUserImportJobs Text

The user pool ID for the user pool that the users are being imported into.

luijMaxResults :: Lens' ListUserImportJobs Natural

The maximum number of import jobs you want the request to return.

Destructuring the Response

listUserImportJobsResponse

Arguments

:: Int
luijrsResponseStatus

-> ListUserImportJobsResponse

Creates a value of ListUserImportJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luijrsPaginationToken - An identifier that can be used to return the next set of user import jobs in the list.

luijrsUserImportJobs - The user import jobs.

luijrsResponseStatus - -- | The response status code.

data ListUserImportJobsResponse

Represents the response from the server to the request to list the user import jobs.

See: listUserImportJobsResponse smart constructor.

Response Lenses

luijrsPaginationToken :: Lens' ListUserImportJobsResponse (Maybe Text)

An identifier that can be used to return the next set of user import jobs in the list.

luijrsUserImportJobs :: Lens' ListUserImportJobsResponse (Maybe (NonEmpty UserImportJobType))

The user import jobs.

luijrsResponseStatus :: Lens' ListUserImportJobsResponse Int

- | The response status code.