Network.AWS.CognitoIdentityProvider.ListUsers

Description

Lists the users in the Amazon Cognito user pool.

Creating a Request

listUsers

Arguments

:: Text
luUserPoolId

-> ListUsers

Creates a value of ListUsers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luPaginationToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luAttributesToGet - An array of strings, where each string is the name of a user attribute to be returned for each user in the search results. If the array is empty, all attributes are returned.

luLimit - Maximum number of users to be returned.

luFilter - A filter string of the form "AttributeName Filter-Type "AttributeValue "". Quotation marks within the filter string must be escaped using the backslash () character. For example, "family_name = "Reddy"". * AttributeName : The name of the attribute to search for. You can only search for one attribute at a time. * Filter-Type : For an exact match, use =, for example, "given_name = "Jon"". For a prefix ("starts with") match, use ^=, for example, "given_name ^= "Jon"". * AttributeValue : The attribute value that must be matched for each user. If the filter string is empty, ListUsers returns all users in the user pool. You can only search for the following standard attributes: * username (case-sensitive) * email * phone_number * name * given_name * family_name * preferred_username * cognito:user_status (called Enabled in the Console) (case-sensitive) * status (case-insensitive) Custom attributes are not searchable. For more information, see Searching for Users Using the ListUsers API and Examples of Using the ListUsers API in the Amazon Cognito Developer Guide .

luUserPoolId - The user pool ID for the user pool on which the search should be performed.

data ListUsers

Represents the request to list users.

See: listUsers smart constructor.

Request Lenses

luPaginationToken :: Lens' ListUsers (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luAttributesToGet :: Lens' ListUsers [Text]

An array of strings, where each string is the name of a user attribute to be returned for each user in the search results. If the array is empty, all attributes are returned.

luLimit :: Lens' ListUsers (Maybe Natural)

Maximum number of users to be returned.

luFilter :: Lens' ListUsers (Maybe Text)

A filter string of the form "AttributeName Filter-Type "AttributeValue "". Quotation marks within the filter string must be escaped using the backslash () character. For example, "family_name = "Reddy"". * AttributeName : The name of the attribute to search for. You can only search for one attribute at a time. * Filter-Type : For an exact match, use =, for example, "given_name = "Jon"". For a prefix ("starts with") match, use ^=, for example, "given_name ^= "Jon"". * AttributeValue : The attribute value that must be matched for each user. If the filter string is empty, ListUsers returns all users in the user pool. You can only search for the following standard attributes: * username (case-sensitive) * email * phone_number * name * given_name * family_name * preferred_username * cognito:user_status (called Enabled in the Console) (case-sensitive) * status (case-insensitive) Custom attributes are not searchable. For more information, see Searching for Users Using the ListUsers API and Examples of Using the ListUsers API in the Amazon Cognito Developer Guide .

luUserPoolId :: Lens' ListUsers Text

The user pool ID for the user pool on which the search should be performed.

Destructuring the Response

listUsersResponse

Arguments

:: Int
lursResponseStatus

-> ListUsersResponse

Creates a value of ListUsersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lursPaginationToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lursUsers - The users returned in the request to list users.

lursResponseStatus - -- | The response status code.

data ListUsersResponse

The response from the request to list users.

See: listUsersResponse smart constructor.

Response Lenses

lursPaginationToken :: Lens' ListUsersResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

lursUsers :: Lens' ListUsersResponse [UserType]

The users returned in the request to list users.

lursResponseStatus :: Lens' ListUsersResponse Int

- | The response status code.