Network.AWS.CognitoIdentityProvider.ListUsersInGroup

Description

Lists the users in the specified group.

Requires developer credentials.

Creating a Request

listUsersInGroup

Arguments

:: Text
luigUserPoolId

-> Text
luigGroupName

-> ListUsersInGroup

Creates a value of ListUsersInGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luigNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luigLimit - The limit of the request to list users.

luigUserPoolId - The user pool ID for the user pool.

luigGroupName - The name of the group.

data ListUsersInGroup

See: listUsersInGroup smart constructor.

Request Lenses

luigNextToken :: Lens' ListUsersInGroup (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luigLimit :: Lens' ListUsersInGroup (Maybe Natural)

The limit of the request to list users.

luigUserPoolId :: Lens' ListUsersInGroup Text

The user pool ID for the user pool.

luigGroupName :: Lens' ListUsersInGroup Text

The name of the group.

Destructuring the Response

listUsersInGroupResponse

Arguments

:: Int
luigrsResponseStatus

-> ListUsersInGroupResponse

Creates a value of ListUsersInGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

luigrsUsers - The users returned in the request to list users.

luigrsNextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luigrsResponseStatus - -- | The response status code.

data ListUsersInGroupResponse

See: listUsersInGroupResponse smart constructor.

Response Lenses

luigrsUsers :: Lens' ListUsersInGroupResponse [UserType]

The users returned in the request to list users.

luigrsNextToken :: Lens' ListUsersInGroupResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

luigrsResponseStatus :: Lens' ListUsersInGroupResponse Int

- | The response status code.