Network.AWS.CognitoIdentityProvider.ResendConfirmationCode

Description

Resends the confirmation (for confirmation of registration) to a specific user in the user pool.

Creating a Request

resendConfirmationCode

Arguments

:: Text
rccClientId

-> Text
rccUsername

-> ResendConfirmationCode

Creates a value of ResendConfirmationCode with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rccSecretHash - A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

rccClientId - The ID of the client associated with the user pool.

rccUsername - The user name of the user to whom you wish to resend a confirmation code.

data ResendConfirmationCode

Represents the request to resend the confirmation code.

See: resendConfirmationCode smart constructor.

Request Lenses

rccSecretHash :: Lens' ResendConfirmationCode (Maybe Text)

A keyed-hash message authentication code (HMAC) calculated using the secret key of a user pool client and username plus the client ID in the message.

rccClientId :: Lens' ResendConfirmationCode Text

The ID of the client associated with the user pool.

rccUsername :: Lens' ResendConfirmationCode Text

The user name of the user to whom you wish to resend a confirmation code.

Destructuring the Response

resendConfirmationCodeResponse

Arguments

:: Int
rccrsResponseStatus

-> ResendConfirmationCodeResponse

Creates a value of ResendConfirmationCodeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rccrsCodeDeliveryDetails - The code delivery details returned by the server in response to the request to resend the confirmation code.

rccrsResponseStatus - -- | The response status code.

data ResendConfirmationCodeResponse

The response from the server when the Amazon Cognito Your User Pools service makes the request to resend a confirmation code.

See: resendConfirmationCodeResponse smart constructor.

Response Lenses

rccrsCodeDeliveryDetails :: Lens' ResendConfirmationCodeResponse (Maybe CodeDeliveryDetailsType)

The code delivery details returned by the server in response to the request to resend the confirmation code.

rccrsResponseStatus :: Lens' ResendConfirmationCodeResponse Int

- | The response status code.