Network.AWS.CognitoIdentityProvider.RespondToAuthChallenge

Description

Responds to the authentication challenge.

Creating a Request

respondToAuthChallenge

Arguments

:: Text
rtacClientId

-> ChallengeNameType
rtacChallengeName

-> RespondToAuthChallenge

Creates a value of RespondToAuthChallenge with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtacChallengeResponses - The challenge responses. These are inputs corresponding to the value of ChallengeName , for example: * SMS_MFA : SMS_MFA_CODE , USERNAME , SECRET_HASH (if app client is configured with client secret). * PASSWORD_VERIFIER : PASSWORD_CLAIM_SIGNATURE , PASSWORD_CLAIM_SECRET_BLOCK , TIMESTAMP , USERNAME , SECRET_HASH (if app client is configured with client secret). * NEW_PASSWORD_REQUIRED : NEW_PASSWORD , any other required attributes, USERNAME , SECRET_HASH (if app client is configured with client secret).

rtacSession - The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

rtacClientId - The app client ID.

rtacChallengeName - The challenge name. For more information, see InitiateAuth . ADMIN_NO_SRP_AUTH is not a valid value.

data RespondToAuthChallenge

The request to respond to an authentication challenge.

See: respondToAuthChallenge smart constructor.

Request Lenses

rtacChallengeResponses :: Lens' RespondToAuthChallenge (HashMap Text Text)

The challenge responses. These are inputs corresponding to the value of ChallengeName , for example: * SMS_MFA : SMS_MFA_CODE , USERNAME , SECRET_HASH (if app client is configured with client secret). * PASSWORD_VERIFIER : PASSWORD_CLAIM_SIGNATURE , PASSWORD_CLAIM_SECRET_BLOCK , TIMESTAMP , USERNAME , SECRET_HASH (if app client is configured with client secret). * NEW_PASSWORD_REQUIRED : NEW_PASSWORD , any other required attributes, USERNAME , SECRET_HASH (if app client is configured with client secret).

rtacSession :: Lens' RespondToAuthChallenge (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

rtacClientId :: Lens' RespondToAuthChallenge Text

The app client ID.

rtacChallengeName :: Lens' RespondToAuthChallenge ChallengeNameType

The challenge name. For more information, see InitiateAuth . ADMIN_NO_SRP_AUTH is not a valid value.

Destructuring the Response

respondToAuthChallengeResponse

Arguments

:: Int
rtacrsResponseStatus

-> RespondToAuthChallengeResponse

Creates a value of RespondToAuthChallengeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtacrsChallengeName - The challenge name. For more information, see InitiateAuth .

rtacrsChallengeParameters - The challenge parameters. For more information, see InitiateAuth .

rtacrsAuthenticationResult - The result returned by the server in response to the request to respond to the authentication challenge.

rtacrsSession - The session which should be passed both ways in challenge-response calls to the service. If the InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

rtacrsResponseStatus - -- | The response status code.

data RespondToAuthChallengeResponse

The response to respond to the authentication challenge.

See: respondToAuthChallengeResponse smart constructor.

Response Lenses

rtacrsChallengeName :: Lens' RespondToAuthChallengeResponse (Maybe ChallengeNameType)

The challenge name. For more information, see InitiateAuth .

rtacrsChallengeParameters :: Lens' RespondToAuthChallengeResponse (HashMap Text Text)

The challenge parameters. For more information, see InitiateAuth .

rtacrsAuthenticationResult :: Lens' RespondToAuthChallengeResponse (Maybe AuthenticationResultType)

The result returned by the server in response to the request to respond to the authentication challenge.

rtacrsSession :: Lens' RespondToAuthChallengeResponse (Maybe Text)

The session which should be passed both ways in challenge-response calls to the service. If the InitiateAuth or RespondToAuthChallenge API call determines that the caller needs to go through another challenge, they return a session with other challenge parameters. This session should be passed as it is to the next RespondToAuthChallenge API call.

rtacrsResponseStatus :: Lens' RespondToAuthChallengeResponse Int

- | The response status code.