Network.AWS.CognitoIdentityProvider.StopUserImportJob

Description

Stops the user import job.

Creating a Request

stopUserImportJob

Arguments

:: Text
sUserPoolId

-> Text
sJobId

-> StopUserImportJob

Creates a value of StopUserImportJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sUserPoolId - The user pool ID for the user pool that the users are being imported into.

sJobId - The job ID for the user import job.

data StopUserImportJob

Represents the request to stop the user import job.

See: stopUserImportJob smart constructor.

Request Lenses

sUserPoolId :: Lens' StopUserImportJob Text

The user pool ID for the user pool that the users are being imported into.

sJobId :: Lens' StopUserImportJob Text

The job ID for the user import job.

Destructuring the Response

stopUserImportJobResponse

Arguments

:: Int
srsResponseStatus

-> StopUserImportJobResponse

Creates a value of StopUserImportJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsUserImportJob - The job object that represents the user import job.

srsResponseStatus - -- | The response status code.

data StopUserImportJobResponse

Represents the response from the server to the request to stop the user import job.

See: stopUserImportJobResponse smart constructor.

Response Lenses

srsUserImportJob :: Lens' StopUserImportJobResponse (Maybe UserImportJobType)

The job object that represents the user import job.

srsResponseStatus :: Lens' StopUserImportJobResponse Int

- | The response status code.