Network.AWS.CognitoIdentityProvider.UpdateUserAttributes

Description

Allows a user to update a specific attribute (one at a time).

Creating a Request

updateUserAttributes

Arguments

:: Text
uuaAccessToken

-> UpdateUserAttributes

Creates a value of UpdateUserAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uuaUserAttributes - An array of name-value pairs representing user attributes. For custom attributes, you must prepend the custom: prefix to the attribute name.

uuaAccessToken - The access token for the request to update user attributes.

data UpdateUserAttributes

Represents the request to update user attributes.

See: updateUserAttributes smart constructor.

Request Lenses

uuaUserAttributes :: Lens' UpdateUserAttributes [AttributeType]

An array of name-value pairs representing user attributes. For custom attributes, you must prepend the custom: prefix to the attribute name.

uuaAccessToken :: Lens' UpdateUserAttributes Text

The access token for the request to update user attributes.

Destructuring the Response

updateUserAttributesResponse

Arguments

:: Int
uuarsResponseStatus

-> UpdateUserAttributesResponse

Creates a value of UpdateUserAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uuarsCodeDeliveryDetailsList - The code delivery details list from the server for the request to update user attributes.

uuarsResponseStatus - -- | The response status code.

data UpdateUserAttributesResponse

Represents the response from the server for the request to update user attributes.

See: updateUserAttributesResponse smart constructor.

Response Lenses

uuarsCodeDeliveryDetailsList :: Lens' UpdateUserAttributesResponse [CodeDeliveryDetailsType]

The code delivery details list from the server for the request to update user attributes.

uuarsResponseStatus :: Lens' UpdateUserAttributesResponse Int

- | The response status code.