Network.AWS.CognitoIdentityProvider.UpdateUserPool

Description

Updates the specified user pool with the specified attributes.

Creating a Request

updateUserPool

Arguments

:: Text
uupUserPoolId

-> UpdateUserPool

Creates a value of UpdateUserPool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uupUserPoolTags - The cost allocation tags for the user pool. For more information, see Adding Cost Allocation Tags to Your User Pool

uupVerificationMessageTemplate - The template for verification messages.

uupEmailVerificationMessage - The contents of the email verification message.

uupSmsAuthenticationMessage - The contents of the SMS authentication message.

uupEmailVerificationSubject - The subject of the email verification message.

uupEmailConfiguration - Email configuration.

uupSmsVerificationMessage - A container with information about the SMS verification message.

uupMFAConfiguration - Can be one of the following values: * OFF - MFA tokens are not required and cannot be specified during user registration. * ON - MFA tokens are required for all user registrations. You can only specify required when you are initially creating a user pool. * OPTIONAL - Users have the option when registering to create an MFA token.

uupLambdaConfig - The AWS Lambda configuration information from the request to update the user pool.

uupSmsConfiguration - SMS configuration.

uupAdminCreateUserConfig - The configuration for AdminCreateUser requests.

uupDeviceConfiguration - Device configuration.

uupAutoVerifiedAttributes - The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

uupPolicies - A container with the policies you wish to update in a user pool.

uupUserPoolId - The user pool ID for the user pool you want to update.

data UpdateUserPool

Represents the request to update the user pool.

See: updateUserPool smart constructor.

Request Lenses

uupUserPoolTags :: Lens' UpdateUserPool (HashMap Text Text)

The cost allocation tags for the user pool. For more information, see Adding Cost Allocation Tags to Your User Pool

uupVerificationMessageTemplate :: Lens' UpdateUserPool (Maybe VerificationMessageTemplateType)

The template for verification messages.

uupEmailVerificationMessage :: Lens' UpdateUserPool (Maybe Text)

The contents of the email verification message.

uupSmsAuthenticationMessage :: Lens' UpdateUserPool (Maybe Text)

The contents of the SMS authentication message.

uupEmailVerificationSubject :: Lens' UpdateUserPool (Maybe Text)

The subject of the email verification message.

uupEmailConfiguration :: Lens' UpdateUserPool (Maybe EmailConfigurationType)

Email configuration.

uupSmsVerificationMessage :: Lens' UpdateUserPool (Maybe Text)

A container with information about the SMS verification message.

uupMFAConfiguration :: Lens' UpdateUserPool (Maybe UserPoolMFAType)

Can be one of the following values: * OFF - MFA tokens are not required and cannot be specified during user registration. * ON - MFA tokens are required for all user registrations. You can only specify required when you are initially creating a user pool. * OPTIONAL - Users have the option when registering to create an MFA token.

uupLambdaConfig :: Lens' UpdateUserPool (Maybe LambdaConfigType)

The AWS Lambda configuration information from the request to update the user pool.

uupSmsConfiguration :: Lens' UpdateUserPool (Maybe SmsConfigurationType)

SMS configuration.

uupAdminCreateUserConfig :: Lens' UpdateUserPool (Maybe AdminCreateUserConfigType)

The configuration for AdminCreateUser requests.

uupDeviceConfiguration :: Lens' UpdateUserPool (Maybe DeviceConfigurationType)

Device configuration.

uupAutoVerifiedAttributes :: Lens' UpdateUserPool [VerifiedAttributeType]

The attributes that are automatically verified when the Amazon Cognito service makes a request to update user pools.

uupPolicies :: Lens' UpdateUserPool (Maybe UserPoolPolicyType)

A container with the policies you wish to update in a user pool.

uupUserPoolId :: Lens' UpdateUserPool Text

The user pool ID for the user pool you want to update.

Destructuring the Response

updateUserPoolResponse

Arguments

:: Int
uuprsResponseStatus

-> UpdateUserPoolResponse

Creates a value of UpdateUserPoolResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uuprsResponseStatus - -- | The response status code.

data UpdateUserPoolResponse

Represents the response from the server when you make a request to update the user pool.

See: updateUserPoolResponse smart constructor.

Response Lenses

uuprsResponseStatus :: Lens' UpdateUserPoolResponse Int

- | The response status code.