Network.AWS.CognitoIdentityProvider.UpdateUserPoolClient

Description

Allows the developer to update the specified user pool client and password policy.

Creating a Request

updateUserPoolClient

Arguments

:: Text
uupcUserPoolId

-> Text
uupcClientId

-> UpdateUserPoolClient

Creates a value of UpdateUserPoolClient with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uupcRefreshTokenValidity - The time limit, in days, after which the refresh token is no longer valid and cannot be used.

uupcExplicitAuthFlows - Explicit authentication flows.

uupcSupportedIdentityProviders - A list of provider names for the identity providers that are supported on this client.

uupcLogoutURLs - A list of allowed logout URLs for the identity providers.

uupcAllowedOAuthFlowsUserPoolClient - Set to TRUE if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

uupcDefaultRedirectURI - The default redirect URI. Must be in the CallbackURLs list.

uupcWriteAttributes - The writeable attributes of the user pool.

uupcReadAttributes - The read-only attributes of the user pool.

uupcAllowedOAuthScopes - A list of allowed OAuth scopes. Currently supported values are "phone" , "email" , "openid" , and Cognito .

uupcAllowedOAuthFlows - Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint. Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

uupcClientName - The client name from the update user pool client request.

uupcCallbackURLs - A list of allowed callback URLs for the identity providers.

uupcUserPoolId - The user pool ID for the user pool where you want to update the user pool client.

uupcClientId - The ID of the client associated with the user pool.

data UpdateUserPoolClient

Represents the request to update the user pool client.

See: updateUserPoolClient smart constructor.

Request Lenses

uupcRefreshTokenValidity :: Lens' UpdateUserPoolClient (Maybe Natural)

The time limit, in days, after which the refresh token is no longer valid and cannot be used.

uupcExplicitAuthFlows :: Lens' UpdateUserPoolClient [ExplicitAuthFlowsType]

Explicit authentication flows.

uupcSupportedIdentityProviders :: Lens' UpdateUserPoolClient [Text]

A list of provider names for the identity providers that are supported on this client.

uupcLogoutURLs :: Lens' UpdateUserPoolClient [Text]

A list of allowed logout URLs for the identity providers.

uupcAllowedOAuthFlowsUserPoolClient :: Lens' UpdateUserPoolClient (Maybe Bool)

Set to TRUE if the client is allowed to follow the OAuth protocol when interacting with Cognito user pools.

uupcDefaultRedirectURI :: Lens' UpdateUserPoolClient (Maybe Text)

The default redirect URI. Must be in the CallbackURLs list.

uupcWriteAttributes :: Lens' UpdateUserPoolClient [Text]

The writeable attributes of the user pool.

uupcReadAttributes :: Lens' UpdateUserPoolClient [Text]

The read-only attributes of the user pool.

uupcAllowedOAuthScopes :: Lens' UpdateUserPoolClient [Text]

A list of allowed OAuth scopes. Currently supported values are "phone" , "email" , "openid" , and Cognito .

uupcAllowedOAuthFlows :: Lens' UpdateUserPoolClient [OAuthFlowType]

Set to code to initiate a code grant flow, which provides an authorization code as the response. This code can be exchanged for access tokens with the token endpoint. Set to token to specify that the client should get the access token (and, optionally, ID token, based on scopes) directly.

uupcClientName :: Lens' UpdateUserPoolClient (Maybe Text)

The client name from the update user pool client request.

uupcCallbackURLs :: Lens' UpdateUserPoolClient [Text]

A list of allowed callback URLs for the identity providers.

uupcUserPoolId :: Lens' UpdateUserPoolClient Text

The user pool ID for the user pool where you want to update the user pool client.

uupcClientId :: Lens' UpdateUserPoolClient Text

The ID of the client associated with the user pool.

Destructuring the Response

updateUserPoolClientResponse

Arguments

:: Int
uupcrsResponseStatus

-> UpdateUserPoolClientResponse

Creates a value of UpdateUserPoolClientResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uupcrsUserPoolClient - The user pool client value from the response from the server when an update user pool client request is made.

uupcrsResponseStatus - -- | The response status code.

data UpdateUserPoolClientResponse

Represents the response from the server to the request to update the user pool client.

See: updateUserPoolClientResponse smart constructor.

Response Lenses

uupcrsUserPoolClient :: Lens' UpdateUserPoolClientResponse (Maybe UserPoolClientType)

The user pool client value from the response from the server when an update user pool client request is made.

uupcrsResponseStatus :: Lens' UpdateUserPoolClientResponse Int

- | The response status code.