Network.AWS.CognitoIdentityProvider.VerifyUserAttribute

Description

Verifies the specified user attributes in the user pool.

Creating a Request

verifyUserAttribute

Arguments

:: Text
vuaAccessToken

-> Text
vuaAttributeName

-> Text
vuaCode

-> VerifyUserAttribute

Creates a value of VerifyUserAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vuaAccessToken - Represents the access token of the request to verify user attributes.

vuaAttributeName - The attribute name in the request to verify user attributes.

vuaCode - The verification code in the request to verify user attributes.

data VerifyUserAttribute

Represents the request to verify user attributes.

See: verifyUserAttribute smart constructor.

Request Lenses

vuaAccessToken :: Lens' VerifyUserAttribute Text

Represents the access token of the request to verify user attributes.

vuaAttributeName :: Lens' VerifyUserAttribute Text

The attribute name in the request to verify user attributes.

vuaCode :: Lens' VerifyUserAttribute Text

The verification code in the request to verify user attributes.

Destructuring the Response

verifyUserAttributeResponse

Arguments

:: Int
vuarsResponseStatus

-> VerifyUserAttributeResponse

Creates a value of VerifyUserAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vuarsResponseStatus - -- | The response status code.

data VerifyUserAttributeResponse

A container representing the response from the server from the request to verify user attributes.

See: verifyUserAttributeResponse smart constructor.

Response Lenses

vuarsResponseStatus :: Lens' VerifyUserAttributeResponse Int

- | The response status code.